+++ src/applications/vaos/appointment-list/components/cards/confirmed/AppointmentDateTime.jsx
@@ -8,7 +8,7 @@ export default function AppointmentDateTime({
   appointmentDate,
   timezone,
   facilityId,
-  isAtlas,
+  twoLineFormat,
 }) {
   if (!appointmentDate.isValid()) {
     return null;
@@ -16,31 +16,22 @@ export default function AppointmentDateTime({
 
   return (
     <>
-      {!isAtlas && (
-        <>
-          {appointmentDate.format('dddd, MMMM D, YYYY [at] h:mm a')}{' '}
-          <span aria-hidden="true">
-            {getAppointmentTimezoneAbbreviation(timezone, facilityId)}
-          </span>
-          <span className="sr-only">
-            {' '}
-            {getAppointmentTimezoneDescription(timezone, facilityId)}
-          </span>
-        </>
+      {!twoLineFormat && (
+        <>{appointmentDate.format('dddd, MMMM D, YYYY [at] h:mm a')} </>
       )}
-      {isAtlas && (
+      {twoLineFormat && (
         <>
           {appointmentDate.format('dddd, MMMM D, YYYY')} <br />
           {appointmentDate.format('h:mm a')}{' '}
-          <span aria-hidden="true">
-            {getAppointmentTimezoneAbbreviation(timezone, facilityId)}
-          </span>
-          <span className="sr-only">
-            {' '}
-            {getAppointmentTimezoneDescription(timezone, facilityId)}
-          </span>
         </>
       )}
+      <span aria-hidden="true">
+        {getAppointmentTimezoneAbbreviation(timezone, facilityId)}
+      </span>
+      <span className="sr-only">
+        {' '}
+        {getAppointmentTimezoneDescription(timezone, facilityId)}
+      </span>
     </>
   );
 }
