--- src/applications/vaos/appointment-list/components/cards/confirmed/VideoAppointmentDateTime.jsx
@@ -1,30 +1,0 @@
-import React from 'react';
-import {
-  getAppointmentTimezoneAbbreviation,
-  getAppointmentTimezoneDescription,
-} from '../../../../utils/appointment';
-
-export default function AppointmentDateTime({
-  appointmentDate,
-  timezone,
-  facilityId,
-}) {
-  if (!appointmentDate.isValid()) {
-    return null;
-  }
-
-  return (
-    <>
-      {appointmentDate.format('dddd, MMMM D, YYYY')}
-      <br />
-      {appointmentDate.format('h:mm a')}{' '}
-      <span aria-hidden="true">
-        {getAppointmentTimezoneAbbreviation(timezone, facilityId)}
-      </span>
-      <span className="sr-only">
-        {' '}
-        {getAppointmentTimezoneDescription(timezone, facilityId)}
-      </span>
-    </>
-  );
-}
