+++ src/applications/vaos/appointment-list/components/cards/confirmed/VideoLink.jsx
@@ -1,0 +1,77 @@
+import React from 'react';
+import classNames from 'classnames';
+import moment from 'moment';
+import {
+  getVideoKind,
+  isVideoGFE,
+  isAtlasLocation,
+} from '../../../../services/appointment';
+import { VIDEO_TYPES } from '../../../../utils/constants';
+
+export default function VideoLink({ appointment }) {
+  const videoKind = getVideoKind(appointment);
+  const isAtlas = isAtlasLocation(appointment);
+  if (isVideoGFE(appointment)) {
+    return (
+      <span>Join the video session from the device provided by the VA.</span>
+    );
+  } else if (videoKind === VIDEO_TYPES.clinic) {
+    return (
+      <span>
+        You must join this video meeting from the VA location listed below.
+      </span>
+    );
+  } else if (isAtlas) {
+    return (
+      <span>
+        You must join this video meeting from the ATLAS (non-VA) location listed
+        below.
+      </span>
+    );
+  } else if (
+    appointment.contained?.[0]?.telecom?.find(tele => tele.system === 'url')
+      ?.value
+  ) {
+    const url = appointment.contained?.[0]?.telecom?.[0]?.value;
+    const diff = moment().diff(moment(appointment.start), 'minutes');
+
+    // Button is enabled 30 minutes prior to start time, until 4 hours after start time
+    const disableVideoLink = diff < -30 || diff > 240;
+    const linkClasses = classNames(
+      'usa-button',
+      'vads-u-margin-left--0',
+      'vads-u-margin-right--1p5',
+      { 'usa-button-disabled': disableVideoLink },
+    );
+
+    return (
+      <div className="vaos-appts__video-visit">
+        <a
+          aria-describedby={
+            disableVideoLink
+              ? `description-join-link-${appointment.id}`
+              : undefined
+          }
+          aria-disabled={disableVideoLink ? 'true' : 'false'}
+          href={url}
+          target="_blank"
+          rel="noopener noreferrer"
+          className={linkClasses}
+          onClick={disableVideoLink ? e => e.preventDefault() : undefined}
+        >
+          Join session
+        </a>
+        {disableVideoLink && (
+          <span
+            id={`description-join-link-${appointment.id}`}
+            className="vads-u-display--block vads-u-font-style--italic"
+          >
+            You can join VA Video Connect 30 minutes prior to the start time
+          </span>
+        )}
+      </div>
+    );
+  }
+
+  return <span>Video visit link unavailable</span>;
+}
